requirements.txt
anyio==3.6.1; python_version >= "3.7" and python_full_version >= "3.6.2"
beautifulsoup4==4.11.1; python_full_version >= "3.6.0"
bs4==0.0.1
certifi==2022.6.15; python_version >= "3.7"
click==8.1.3; python_version >= "3.7"
colorama==0.4.5; python_version >= "3.7" and python_full_version < "3.0.0" and platform_system == "Windows" or platform_system == "Windows" and python_version >= "3.7" and python_full_version >= "3.5.0"
fastapi==0.79.0; python_full_version >= "3.6.1"
gunicorn==20.1.0; python_version >= "3.5"
h11==0.12.0; python_version >= "3.7"
httpcore==0.15.0; python_version >= "3.7"
httpx==0.23.0; python_version >= "3.7"
idna==3.3; python_version >= "3.7" and python_full_version >= "3.6.2"
pydantic==1.9.1; python_full_version >= "3.6.1"
python-dotenv==0.20.0; python_version >= "3.5"
rfc3986==1.5.0; python_version >= "3.7"
sniffio==1.2.0; python_version >= "3.7" and python_full_version >= "3.6.2"
soupsieve==2.3.2.post1; python_version >= "3.6" and python_full_version >= "3.6.0"
starlette==0.19.1; python_version >= "3.6" and python_full_version >= "3.6.1"
typing-extensions==4.3.0; python_version >= "3.7" and python_full_version >= "3.6.1" and python_version < "3.10"
uvicorn==0.18.2; python_version >= "3.7"
